pyproject.toml
[project]
name = "code-context-builder"
version = "0.1.0"
description = "Create an agent-friendly index of your codebase"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "tree-sitter>=0.21.0",
    "tree-sitter-languages",
    "httpx",
    "python-dotenv",
    "click",
    "rich",
    "pathspec"
]
requires-python = ">=3.10"

[project.scripts]
ccb = "ccb.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
